Move ressources to new file
