+++ tests/geo_ressources.py
@@ -1,0 +1,21 @@
+from haversine import Unit
+
+LYON = (45.7597, 4.8422)
+PARIS = (48.8567, 2.3508)
+NEW_YORK = (40.7033962, -74.2351462)
+LONDON = (51.509865, -0.118092)
+
+EXPECTED_LYON_PARIS = {Unit.KILOMETERS: 392.2172595594006,
+            Unit.METERS: 392217.2595594006,
+            Unit.MILES: 243.71250609539814,
+            Unit.NAUTICAL_MILES: 211.78037755311516,
+            Unit.FEET: 1286802.0326751503,
+            Unit.INCHES: 15441624.392102592,
+            Unit.RADIANS: 0.386810597795104,
+            Unit.DEGREES: 22.162614724591847}
+EXPECTED_LYON_NEW_YORK = {Unit.KILOMETERS: 6163.43638211,
+            Unit.METERS: 61634363.8211 }
+EXPECTED_PARIS_NEW_YORK = {Unit.KILOMETERS: 5853.32898662,
+            Unit.METERS: 58533289.8662 }
+EXPECTED_LONDON_PARIS = {Unit.KILOMETERS: 343.37455271}
+EXPECTED_LONDON_NEW_YORK = {Unit.KILOMETERS: 5586.48447423}
