+++ tests/test_haversine.py
@@ -1,21 +1,7 @@
-from haversine import haversine, haversine_vector, Unit
+from haversine import haversine, Unit
 from math import pi
-from numpy.testing import assert_allclose
 
-LYON = (45.7597, 4.8422)
-PARIS = (48.8567, 2.3508)
-LONDON = (51.509865, -0.118092)
-NEW_YORK = (40.7033962, -74.2351462)
-
-EXPECTED_LYON_PARIS = {Unit.KILOMETERS: 392.2172595594006,
-            Unit.METERS: 392217.2595594006,
-            Unit.MILES: 243.71250609539814,
-            Unit.NAUTICAL_MILES: 211.78037755311516,
-            Unit.FEET: 1286802.0326751503,
-            Unit.INCHES: 15441624.392102592,
-            Unit.RADIANS: 0.061562818679421795,
-            Unit.DEGREES: 3.5272896852600164}
-
+from tests.geo_ressources import LYON, PARIS, NEW_YORK, LONDON, EXPECTED_LYON_PARIS
 
 def haversine_test_factory(unit):
     def test():
@@ -41,14 +27,6 @@ def test_units_enum():
     from haversine.haversine import _CONVERSIONS
     assert all(unit in _CONVERSIONS for unit in Unit)
 
-
-def test_haversine_vector_comb():
-    expected = [[ 392.21725956,  343.37455271], [6163.43638211, 5586.48447423]]
-
-    assert_allclose( # See https://numpy.org/doc/stable/reference/generated/numpy.testing.assert_allclose.html#numpy.testing.assert_allclose
-        haversine_vector([LYON, LONDON], [PARIS, NEW_YORK], Unit.KILOMETERS, comb=True),
-        expected
-    )
 
 def test_haversine_deg_rad():
     """
